Remove item update endpoint, add comments
Modified TodoController to remove the item update endpoint, too many error modes and not enough time to handle them properly.  Modified path for delete item endpoint to use list as the parent.  Fixed list update endpoint to return 404 if there's no such list to update.  Added javadoc.

+++ src/main/java/org/molodowitch/cs90/todos/controller/TodoController.java
@@ -1,6 +1,5 @@
 package org.molodowitch.cs90.todos.controller;
 
-import org.molodowitch.cs90.todos.model.TodoItem;
 import org.molodowitch.cs90.todos.model.TodoList;
 import org.molodowitch.cs90.todos.services.TodoService;
 import org.slf4j.Logger;
@@ -15,6 +14,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * REST controller providing endpoints for a To-do application back end.
+ * Note that this was built as a demo to investigate AWS Serverless.  It's not
+ * intended for true production use, as it skips some potential error modes
+ * and doesn't provide the full API that would be required for an actual
+ * To-do application.
+ */
 @RestController
 @EnableWebMvc
 public class TodoController {
@@ -27,6 +33,13 @@ public class TodoController {
         this.todoService = todoService;
     }
 
+    /**
+     * Ping endpoint created by Maven archetype, left in to serve as database-independent
+     * test endpoint.
+     * Does no actual logic, simply returns a response when reached.
+     *
+     * @return Map matching "pong" to "Hello, World!"
+     */
     @GetMapping(path = "/ping")
     public Map<String, String> ping() {
         Map<String, String> pong = new HashMap<>();
@@ -34,39 +47,76 @@ public class TodoController {
         return pong;
     }
 
+    /**
+     * Gets all saved to-do lists, including the items on the lists.
+     *
+     * @return List of TodoList objects, each with all to-do items loaded
+     */
     @GetMapping(path = "/lists")
     public List<TodoList> getTodoLists() {
         return todoService.getAllLists();
     }
 
+    /**
+     * Gets the to-do list with the specified persistence index, if any.
+     *
+     * @param listId persistence index of the to-do list to retrieve
+     * @return Optional containing the TodoList with loaded to-do items for the list with the
+     *         specified persistence index, or an empty Optional if there's no such list
+     */
     @GetMapping(path = "/lists/{id}")
     public ResponseEntity<TodoList> getTodoList(@PathVariable("id") long listId) {
         return ResponseEntity.of(todoService.getList(listId));
     }
 
+    /**
+     * Creates a new to-do list entry based on the specified data.
+     *
+     * @param list TodoList object parsed from the request containing the new list information
+     * @return saved TodoList data, including persistence index
+     */
     @PostMapping(path = "/lists")
     public TodoList createNewList(@RequestBody TodoList list) {
         return todoService.saveList(list);
     }
 
+    /**
+     * Modifies the specified to-do list entry, updating it with the data in the TodoList
+     * parsed from the request.  Note that this will overwrite any previous values, including
+     * to-do list items.  Use this method to modify existing lists, including adding or
+     * editing to-do items.
+     *
+     * @param list   TodoList object containing to-do list data parsed from the request body
+     * @param listId persistence index of the to-do list to update
+     * @return updated saved TodoList data
+     */
     @PutMapping(path = "/lists/{id}")
-    public TodoList updateList(@RequestBody TodoList list, @PathVariable("id") long listId) {
-        return todoService.saveList(list);
+    public ResponseEntity<TodoList> updateList(@RequestBody TodoList list, @PathVariable("id") long listId) {
+        if (todoService.doesListExist(listId)) {
+            list.setId(listId);
+            return ResponseEntity.ok(todoService.saveList(list));
+        }
+        else {
+            LOG.error("No list found to update with index {}", listId);
+            return ResponseEntity.notFound().build();
+        }
     }
 
-    @PutMapping(path = "/items/{itemId}")
-    public TodoItem updateItem(@RequestBody TodoItem item, @PathVariable("itemId") long itemId) {
-        return todoService.editItem(item);
-    }
-
-    @DeleteMapping(path = "/items/{itemId}")
+    /**
+     * Deletes the specified to-do item.
+     *
+     * @param itemId persistence index of the to-do item to delete
+     * @return 204 No Content response if item is successfully deleted,
+     *         or 404 Not Found if there's no to-do item with the specified itemId
+     */
+    @DeleteMapping(path = "/lists/{listId}/items/{itemId}")
     public ResponseEntity<String> deleteItem(@PathVariable("itemId") long itemId) {
-        var exists = todoService.deleteItem(itemId);
-        if (exists) {
+        if (todoService.doesItemExist(itemId)) {
+            todoService.deleteItem(itemId);
             return ResponseEntity.noContent().build();
         }
         else {
-            String message = "Todo item " + itemId + " not found";
+            String message = "Couldn't delete to-do item " + itemId + ", item not found";
             LOG.error(message);
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, message);
         }
